% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_sources.R
\name{bind_sources}
\alias{bind_sources}
\title{Row-bind a list of data.frames into one}
\usage{
bind_sources(data, ...)
}
\arguments{
\item{data}{A list of data.frame.}

\item{...}{Variables in the output. The argument name should be the new name in the output, and the right hand side of argument is a character vector of the original names. The name vector and the data will be matched by position. if an output variable only came from some of the sources, fill the name vector to a length equal to the number of sources with NA, e.g., 'var' only come from the second out of three sources, var = c(NA, 'nm_in_src2', NA).}
}
\value{
A data.frame contains combined rows of the input list with variables specified by ...
}
\description{
Row-bind a list of data.frames into one
}
\examples{
df1 <- subset(iris, Species == "setosa")
df2 <- subset(iris, Species == "versicolor")
df3 <- subset(iris, Species == "virginica")

bind_sources(list(df1, df2, df3),
 s_l = "Sepal.Length",
 s_w = "Sepal.Width",
 p_l_setosa = c("Petal.Length", NA, NA),
 p_l_virginica = c(NA, NA, "Petal.Length")
)
}
