% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdb-package.R
\docType{package}
\name{healthdb-package}
\alias{healthdb}
\alias{healthdb-package}
\title{healthdb: Working with Healthcare Databases}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A system for identifying diseases or events from healthcare databases and preparing data for epidemiological studies. It includes capabilities not supported by 'SQL', such as matching strings by 'stringr' style regular expressions, and can compute comorbidity scores (Quan et al. (2005) \doi{10.1097/01.mlr.0000182534.19832.83}) directly on a database server. The implementation is based on 'dbplyr' with full 'tidyverse' compatibility.
}
\section{Package options}{

The behavior of some functions can be adjusted with [options()]:
\describe{
 \item{healthdb.verbose}{Whether functions explain what they are doing and give an overview of the results. The default is TRUE. Use `options(healthdb.verbose = FALSE)` to turn the messages off.}
 \item{healthdb.force_proceed}{Some operations have to download data from the database first, which may be slow. By default (FALSE), functions ask for your confirmation before downloading. Use `options(healthdb.force_proceed = TRUE)` to skip the prompts.}
 \item{healthdb.check_con}{Whether functions test that the database connection is alive (by sending a trivial `SELECT 1` query) before building queries on it. The default is TRUE. Use `options(healthdb.check_con = FALSE)` to skip the test; this saves one trip to the database per step, which can add up in long pipelines, but a dropped connection would then surface as a less informative error.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KevinHzq/healthdb}
  \item \url{https://kevinhzq.github.io/healthdb/}
  \item Report bugs at \url{https://github.com/KevinHzq/healthdb/issues}
}

}
\author{
\strong{Maintainer}: Kevin Hu \email{kevin.hu@bccdc.ca} (\href{https://orcid.org/0000-0003-0254-5277}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kevin Hu \email{kevin.hu@bccdc.ca} (\href{https://orcid.org/0000-0003-0254-5277}{ORCID}) [copyright holder]
}

}
\keyword{internal}
