% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_encounter.R
\name{identify_encounter}
\alias{identify_encounter}
\title{Title}
\usage{
identify_encounter(
  data,
  from_var,
  match_vals,
  match_type = c("in", "start", "regex", "like", "between", "glue_sql"),
  if_all = FALSE,
  verbose = TRUE,
  query_only = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
