% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_any_dates.R
\name{is_any_dates}
\alias{is_any_dates}
\title{Interpret if n dates drawn from a vector could be some days apart within some years}
\usage{
is_any_dates(x, n, apart = NULL, within = NULL, ...)
}
\arguments{
\item{x}{A character or Date vector}

\item{n}{An integer for the size of a draw}

\item{apart}{An integer specifying the minimum gap (in days) between adjacent dates in a draw.}

\item{within}{An integer specifying the maximum time span (in days) of a draw.}

\item{...}{Additional argument passing to as.Date for character to date conversion.}
}
\value{
Logical for whether there is any draw from x satisfied the conditions
}
\description{
Return logical if there could be at least one set of n elements taken from x that adjacent elements in the set are at least certain days apart AND the dates in the set are within the specified time span. When identifying disease from administrative data, definitions often require ,e.g., n diagnoses that are at least some days apart within some years. This function is intended for such use and optimized to avoid looping through all possible n size combinations in x.
}
\examples{
dates_of_records <- sample(seq(as.Date("2015-01-01"), as.Date("2021-12-31"), 7), 10)

# whether there is any 3 records at least 30 days apart within 2 years
is_any_dates(dates_of_records, n = 3, apart = 30, within = 365 * 2)

# specified either apart or within or both
is_any_dates(dates_of_records, n = 2, within = 365)

}
