% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odcfun-package.R
\docType{package}
\name{odcfun-package}
\alias{odcfun}
\alias{odcfun-package}
\title{odcfun: Overdose Cohort Team In-house Functions}
\description{
Designed for identifying disease cases from admin data for epidemiological studies. The implementation focused on code readability and re-usability.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KevinHzq/odcfun}
  \item Report bugs at \url{https://github.com/KevinHzq/odcfun/issues}
}

}
\keyword{internal}
