% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdb-package.R
\docType{package}
\name{healthdb-package}
\alias{healthdb}
\alias{healthdb-package}
\title{healthdb: Tools for Working with Healthcare Database}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A system for identifying diseases or events from healthcare database and preparing data for epidemiological studies. It features abilities that are not supported by SQL database, such as matching strings by 'stringr' style regular expression. The implementation is based on 'dbplyr' providing full compatibility with 'tidyverse'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KevinHzq/healthdb}
  \item \url{https://kevinhzq.github.io/healthdb/}
  \item Report bugs at \url{https://github.com/KevinHzq/healthdb/issues}
}

}
\author{
\strong{Maintainer}: Kevin Hu \email{kevin.hu@bccdc.ca} (\href{https://orcid.org/0000-0003-0254-5277}{ORCID}) [copyright holder]

}
\keyword{internal}
