% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odcfun-package.R
\docType{package}
\name{odcfun-package}
\alias{odcfun}
\alias{odcfun-package}
\title{odcfun: Tools for working with healthcare database}
\description{
A set of tools for identifying diseases or events from healthcare database and preparing data for epidemiological studies. It features abilities that are not natively support by database, such as matching strings by 'stringr' style regular expression and using 'LIKE' operator with multiple patterns in a vector. The implementation is based on 'dbplyr' with focuses on code readability and re-usability. Three types of functions are included: interactive functions – for customizing complex definitions; call building functions – for batch execution of simple definitions; miscellaneous functions – for data wrangling, computing age and comorbidity index, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KevinHzq/odcfun}
  \item \url{https://kevinhzq.github.io/odcfun/}
  \item Report bugs at \url{https://github.com/KevinHzq/odcfun/issues}
}

}
\keyword{internal}
