% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odcfun-package.R
\docType{package}
\name{odcfun-package}
\alias{odcfun}
\alias{odcfun-package}
\title{odcfun: Overdose Cohort Team In-house Functions}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
