% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odcfun-package.R
\docType{package}
\name{odcfun-package}
\alias{odcfun}
\alias{odcfun-package}
\title{odcfun: Using Administrative Database for Epidemiology}
\description{
Designed for identifying disease cases from administrative database for epidemiological studies. The implementation is based on 'dbplyr' with a focuse on code readability and re-usability.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KevinHzq/odcfun}
  \item \url{https://kevinhzq.github.io/odcfun/}
  \item Report bugs at \url{https://github.com/KevinHzq/odcfun/issues}
}

}
\keyword{internal}
